Updating components to a more generic state
