+++ src/js/components/generateDetachedFiles/DetachedFilesFYP.jsx
@@ -25,23 +25,31 @@ const initialPeriod = defaultPeriods(true);
 const propTypes = {
     type: PropTypes.oneOf(['dabs', 'fabs']),
     clickedDownload: PropTypes.func,
-    generateFileA: PropTypes.func,
+    generateFYPFile: PropTypes.func,
     status: PropTypes.string,
     errorType: PropTypes.string,
     errorMessage: PropTypes.string,
-    showDownload: PropTypes.bool
+    showDownload: PropTypes.bool,
+    fileType: PropTypes.string,
+    fileLabel: PropTypes.string,
+    fileTitle: PropTypes.string,
+    description: PropTypes.element
 };
 
 const defaultProps = {
     clickedDownload: null,
-    generateFileA: () => { },
+    generateFYPFile: () => { },
     status: '',
     errorType: '',
     errorMessage: '',
-    showDownload: false
+    showDownload: false,
+    fileType: '',
+    fileLabel: '',
+    fileTitle: '',
+    description: null
 };
 
-export default class DetachedFileA extends React.Component {
+export default class DetachedFilesFYP extends React.Component {
     constructor(props) {
         super(props);
 
@@ -105,7 +113,7 @@ export default class DetachedFileA extends React.Component {
         this.setState({
             downloadFields: `${this.state.agencyName} | FY ${this.state.fy} | ${minPeriod} - ${maxPeriod}`
         });
-        this.props.generateFileA(this.state.agency, this.state.codeType, this.state.period,
+        this.props.generateFYPFile(this.state.agency, this.state.codeType, this.state.period,
             parseInt(this.state.fy, 10));
     }
 
@@ -116,7 +124,7 @@ export default class DetachedFileA extends React.Component {
         }
 
         let submissionLink = null;
-        if (this.props.status === 'done') {
+        if (this.props.status === 'done' && this.props.fileType !== 'BOC') {
             submissionLink = (
                 <div className="submission-link">
                     <Link to="/submissionGuide">
@@ -167,68 +175,12 @@ export default class DetachedFileA extends React.Component {
                                         </div>
                                         <div className="description">
                                             <div className="description__content">
-                                                <p>
-                                                    Select a Fiscal Year, Period, and Agency, and the Broker will
-                                                    generate a provisional File A for that agency. You are responsible
-                                                    for reviewing and amending (e.g., if any TAS need to be added or
-                                                    deleted) the generated File A for accuracy and completeness before
-                                                    certifying to it in any submission.
-                                                </p>
-                                                <p>
-                                                    File A generation changes nothing about the existing DABS
-                                                    submission process (including validation, cross-file validation,
-                                                    and certification), other than facilitating reconciliation and
-                                                    providing a starting point for agency submissions. File A is
-                                                    generated outside the submission context. The Broker will not
-                                                    automatically attach it to any submission.
-                                                </p>
-                                                <p>
-                                                    Consistent with DATA Broker Quarterly Submissions (DABS)
-                                                    guidelines, generated files are at the &apos;agency-wide level&apos;
-                                                    the goal is to include all accounts for a given agency that are
-                                                    appropriate for DATA Act submissions. Financing accounts are
-                                                    automatically excluded, and child allocation accounts are bucketed
-                                                    with the child agency. File A is generated based on GTAS SF-133
-                                                    data, which GTAS provides to the Broker on a daily basis during any
-                                                    reporting/revision windows (note that it will take up to 24
-                                                    hours for changes agencies make in GTAS to be reflected in the
-                                                    Broker). For a more detailed explanation of the approach for
-                                                    generating File A, see the Practices and Procedures document
-                                                    available on the&nbsp;
-                                                    <a
-                                                        target="_blank"
-                                                        rel="noopener noreferrer"
-                                                        href={"https://fiscal.treasury.gov/data-transparency/" +
-                                                            "GSDM-current.html"}>
-                                                        GSDM
-                                                    </a>
-                                                    &nbsp;page of the Data Transparency site of the
-                                                    Bureau of the Fiscal Service.
-                                                </p>
-                                                <p>
-                                                    Note: Periods are available to generate starting on the 1st of the
-                                                    following month (for example, P02 data will be available to generate
-                                                    starting Dec 1). However, until the GTAS window for a given period
-                                                    is complete, File A Data is subject to change and may need to be
-                                                    regenerated in order to reflect the final state of GTAS data after
-                                                    the window closes.
-
-                                                    While Period 01 data is automatically included with data from later
-                                                    periods (because File A Data is cumulative within the Fiscal year),
-                                                    it is not selectable on its own and therefore will not be visible
-                                                    until Dec 1 with Period 02.
-                                                </p>
-                                                <p>
-                                                    Note: The generated file A column GTASStatus contains the status of
-                                                    each TAS in the GTAS system. This data element is for information
-                                                    only. It is not a GSDM data element, and is not required for any
-                                                    File A submissions.
-                                                </p>
+                                                {this.props.description}
                                             </div>
                                         </div>
                                         <div className="file-a-section">
                                             <div className="file-a-section__label">
-                                                File A: Appropriations Accounts
+                                                {this.props.fileTitle}
                                             </div>
 
                                             <div className="file-a-section__date">
@@ -244,7 +196,7 @@ export default class DetachedFileA extends React.Component {
                                         </div>
                                         <DownloadFile
                                             fileInfo={this.state.downloadFields}
-                                            label="File A: Appropriations Accounts"
+                                            label={this.props.fileTitle}
                                             errorType={this.props.errorType}
                                             errorMessage={this.props.errorMessage}
                                             clickedDownload={this.props.clickedDownload}
@@ -253,7 +205,8 @@ export default class DetachedFileA extends React.Component {
                                         <GenerateButton
                                             agency={this.state.agency}
                                             generate={this.generate}
-                                            status={this.props.status} />
+                                            status={this.props.status}
+                                            fileLabel={this.props.fileLabel} />
                                         {submissionLink}
                                     </div>
                                 </div>
@@ -267,5 +220,5 @@ export default class DetachedFileA extends React.Component {
     }
 }
 
-DetachedFileA.propTypes = propTypes;
-DetachedFileA.defaultProps = defaultProps;
+DetachedFilesFYP.propTypes = propTypes;
+DetachedFilesFYP.defaultProps = defaultProps;
